app/requirements.txt
alabaster==0.7.12
amqp==5.0.9
asgiref==3.4.1
astro-ghost==0.1.25
astro-prospector==1.0.0
astropy==5.0.4
astroquery==0.4.4
Babel==2.9.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bokeh==2.4.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
DateTime==4.3
Django==3.2.9
django-celery-beat==2.2.1
django-crispy-forms==1.14.0
django-timezone-field==4.2.3
docutils==0.17.1
flower==1.0.0
fonttools==4.28.2
furo==2021.11.23
h5py==3.6.0
html5lib==1.1
humanize==3.13.1
idna==3.3
imagesize==1.3.0
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.2
install==1.3.5
Jinja2==3.1.1
joblib==1.1.0
joypy==0.2.5
keyring==23.4.0
kiwisolver==1.3.2
kombu==5.2.3
MarkupSafe==2.0.1
matplotlib==3.5.0
mimeparse==0.1.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
photutils==1.2.0
Pillow==8.4.0
pip==21.3.1
prometheus-client==0.13.0
prompt-toolkit==3.0.24
pyerfa==2.0.0.1
Pygments==2.10.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pyvo==1.3
PyYAML==6.0
requests==2.26.0
rfpimp==1.3.7
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools==59.4.0
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
sphinx-copybutton==0.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
typing_extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
zipp==3.6.0
zope.interface==5.4.0
pymysql==1.0.2
mysqlclient
gunicorn==20.1.0
coverage
scikit-image
dustmaps
astro-sedpy
django-revproxy
